recom-auth-scrap/requirements.txt
beautifulsoup4==4.9.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
nltk==3.6.2
mysqlclient==2.1.0
pandas==1.3.4
requests==2.26.0
scikit-learn==1.0.2
